Moving data manipulation helpers into dedicated helper file.
