+++ resources/assets/components/actions/VoterRegistrationAction/VoterRegistrationAction.js
@@ -3,8 +3,8 @@ import gql from 'graphql-tag';
 import PropTypes from 'prop-types';
 
 import Card from '../../utilities/Card/Card';
-import { dynamicString } from '../../../helpers';
 import { getUserId } from '../../../helpers/auth';
+import { dynamicString } from '../../../helpers/data';
 import PrimaryButton from '../../utilities/Button/PrimaryButton';
 import TextContent from '../../utilities/TextContent/TextContent';
 import AnalyticsWaypoint from '../../utilities/AnalyticsWaypoint/AnalyticsWaypoint';
